Restructured Python files into separate folders
