+++ python/pdstools/README.md
@@ -1,0 +1,3 @@
+# Python pdstools
+
+This top-level folder contains all of the subdirectories containing the library.
