+++ python/pdstools/__init__.py
@@ -1,5 +1,5 @@
-from .ADMDatamart import ADMDatamart
-from .cdh_utils import readDSExport
-from .datasets import CDHSample
-from .ADMTrees import ADMTrees, MultiTrees
-from .ValueFinder import ValueFinder
+from .adm.ADMDatamart import ADMDatamart
+from .adm.ADMTrees import ADMTrees, MultiTrees
+from .valuefinder.ValueFinder import ValueFinder
+from .utils import cdh_utils
+from .utils import datasets
