+++ python/pdstools/adm/README.md
@@ -1,0 +1,7 @@
+# ADM
+
+This subdirectory contains all analyses using the ADM Datamart. Currently, that is the ADMDatamart class and the ADMTrees.
+
+The ADMDatamart class in particular also uses the plots in the `plots` folder. 
+
+Both use utilities in the `utils` folder.
