+++ python/pdstools/app/README.md
@@ -1,0 +1,5 @@
+# PDS App
+
+This directory contains the files creating the pdstools app. 
+
+Use the app by running `pdstools run` in the terminal after having installed it through pip.
