+++ python/pdstools/ih/README.md
@@ -1,0 +1,3 @@
+# IH
+
+This directory contains all methods to analyse Interaction History datasets.
