+++ python/pdstools/plots/README.md
@@ -1,0 +1,8 @@
+# Plots
+
+This directory contains all methods used to create the visualisations, mainly used by the `ADMDatamart` class.
+
+It contains of three files:
+- `plot_base.py` takes care of preprocessing
+- `plot_plotly` takes the preprocessed data and creates Plotly charts
+- `plot_mpl.py` takes the preprocessed data and creates matplotlib/seaborn plots
