+++ python/pdstools/utils/README.md
@@ -1,0 +1,3 @@
+# Utils
+
+This directory contains several utility methods, such as importing the sample `datasets`, and general utilities in `cdh_utils.py`.
