+++ python/pdstools/valuefinder/README.md
@@ -1,0 +1,3 @@
+# Value Finder
+
+This directory contains the Value Finder class.
