+++ python/pdstools/valuefinder/ValueFinder.py
@@ -10,7 +10,7 @@ import plotly.figure_factory as ff
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from typing import Optional, Union
-from . import cdh_utils
+from ..utils import cdh_utils
 
 
 class ValueFinder:
